Use single project to target .NET Framework & Core
SDK-style projects allow multi-targeting which makes this much simpler.

The AsysTech.CredentialManager project will also pack a single package that targets multiple frameworks
