+++ CredentialManager.sln
@@ -19,10 +19,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{0609161C-CF5
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{7E19D2AF-C0F0-4229-8E95-5B8ADC0BE732}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdysTech.CredentialManager.Core", "src\AdysTech.CredentialManager.Core\AdysTech.CredentialManager.Core.csproj", "{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CredentialManager.Core.Test", "tests\CredentialManager.Core.Test\CredentialManager.Core.Test.csproj", "{6A048AD7-2768-4A08-893E-DB33032DA79C}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,28 +37,12 @@ Global
 		{ABA77D40-8D43-42EE-BE03-44E4AD93672A}.Release|x64.Build.0 = Release|Any CPU
 		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|x64.ActiveCfg = Debug|x64
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|x64.Build.0 = Debug|x64
+		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|x64.Build.0 = Debug|Any CPU
 		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|x64.ActiveCfg = Release|x64
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|x64.Build.0 = Release|x64
-		{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C}.Debug|x64.Build.0 = Debug|Any CPU
-		{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C}.Release|x64.ActiveCfg = Release|Any CPU
-		{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C}.Release|x64.Build.0 = Release|Any CPU
-		{6A048AD7-2768-4A08-893E-DB33032DA79C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6A048AD7-2768-4A08-893E-DB33032DA79C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6A048AD7-2768-4A08-893E-DB33032DA79C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{6A048AD7-2768-4A08-893E-DB33032DA79C}.Debug|x64.Build.0 = Debug|Any CPU
-		{6A048AD7-2768-4A08-893E-DB33032DA79C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6A048AD7-2768-4A08-893E-DB33032DA79C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6A048AD7-2768-4A08-893E-DB33032DA79C}.Release|x64.ActiveCfg = Release|Any CPU
-		{6A048AD7-2768-4A08-893E-DB33032DA79C}.Release|x64.Build.0 = Release|Any CPU
+		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|x64.ActiveCfg = Release|Any CPU
+		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -70,8 +50,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{ABA77D40-8D43-42EE-BE03-44E4AD93672A} = {0609161C-CF52-44B3-A09E-6CCD4BCCB30F}
 		{5DCA2C07-C1F3-424F-B934-F952595586CE} = {7E19D2AF-C0F0-4229-8E95-5B8ADC0BE732}
-		{B21EEA8C-6EAC-4B2C-A16C-BD58C363549C} = {0609161C-CF52-44B3-A09E-6CCD4BCCB30F}
-		{6A048AD7-2768-4A08-893E-DB33032DA79C} = {7E19D2AF-C0F0-4229-8E95-5B8ADC0BE732}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {36A3FDA1-ACC5-4934-8E8A-DBE3ACADFE39}
