--- src/AdysTech.CredentialManager.Core/AdysTech.CredentialManager.Core.csproj
@@ -1,39 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-	  <PackageId>AdysTech.CredentialManager.Core</PackageId>   
-   <AssemblyName>AdysTech.CredentialManager</AssemblyName>
-    <RootNamespace>AdysTech.CredentialManager</RootNamespace>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageProjectUrl>https://github.com/AdysTech/CredentialManager</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/AdysTech/CredentialManager</RepositoryUrl>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <Copyright>© AdysTech 2016-2020</Copyright>
-    <Authors>AdysTech</Authors>
-    <Version>2.0.0.0</Version>
-    <AssemblyVersion>2.0.0.0</AssemblyVersion> 
-    <PackageReleaseNotes></PackageReleaseNotes> 
-    <Description>C# wrapper around CredWrite / CredRead functions to store and retrieve from Windows Credential Store.
-The library provides
-         1. Prompt user for Windows Generic Credentials
-         2. Save Windows Generic Credentials
-         3. Retrieve saved Windows Generic Credentials
-         4. Add or edit ccomments, store additional attribute objects associated with Credentials
-
-Package supporting .Net Framework 4.5+ (AdysTech.CredentialManager) is available at https://www.nuget.org/packages/AdysTech.CredentialManager</Description> 
-    <FileVersion>2.0.0.0</FileVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="..\Credential.cs" Link="Credential.cs" />
-    <Compile Include="..\CredentialExtensions.cs" Link="CredentialExtensions.cs" />
-    <Compile Include="..\CredentialManager.cs" Link="CredentialManager.cs" />
-    <Compile Include="..\CriticalCredentialHandle.cs" Link="CriticalCredentialHandle.cs" />
-    <Compile Include="..\NativeCode.cs" Link="NativeCode.cs" />
-    <Compile Include="..\ICredential.cs" Link="ICredential.cs" />
-    <Compile Include="..\PublicEnums.cs" Link="PublicEnums.cs" />
-    <Compile Include="..\CredentialAPIException.cs" Link="CredentialAPIException.cs" />
-  </ItemGroup>
-
-</Project>
