--- src/AdysTech.CredentialManager.Core/internalModifier.cs
@@ -1,2 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-[assembly: InternalsVisibleTo("CredentialManagerTest")]
