+++ src/AdysTech.CredentialManager/AdysTech.CredentialManager.csproj
@@ -1,71 +1,27 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{ABA77D40-8D43-42EE-BE03-44E4AD93672A}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>AdysTech.CredentialManager</RootNamespace>
-    <AssemblyName>AdysTech.CredentialManager</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
+    <TargetFrameworks>net45;netstandard2.1</TargetFrameworks>
+    <AssemblyOriginatorKeyFile>Properties\key.snk</AssemblyOriginatorKeyFile>
     <LangVersion>8</LangVersion>
   </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
+
   <PropertyGroup>
-    <SignAssembly>true</SignAssembly>
+    <PackageId>CredentialManager</PackageId>
+    <Version>2.0.0.0</Version>
+    <PackageProjectUrl>https://github.com/AdysTech/CredentialManager</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/AdysTech/CredentialManager</RepositoryUrl>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <Copyright>© AdysTech 2016-2020</Copyright>
+    <Authors>AdysTech</Authors>
+    <Description>C# wrapper around CredWrite / CredRead functions to store and retrieve from Windows Credential Store.
+The library provides
+         1. Prompt user for Windows Generic Credentials
+         2. Save Windows Generic Credentials
+         3. Retrieve saved Windows Generic Credentials
+         4. Add or edit ccomments, store additional attribute objects associated with Credentials
+
+Package supporting .Net Framework 4.5+ (AdysTech.CredentialManager) is available at https://www.nuget.org/packages/AdysTech.CredentialManager</Description> 
   </PropertyGroup>
-  <PropertyGroup>
-    <AssemblyOriginatorKeyFile>Properties\key.snk</AssemblyOriginatorKeyFile>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="..\Credential.cs" />
-    <Compile Include="..\CredentialExtensions.cs">
-      <Link>CredentialExtensions.cs</Link>
-    </Compile>
-    <Compile Include="..\CredentialManager.cs" />
-    <Compile Include="..\NativeCode.cs" />
-    <Compile Include="..\CriticalCredentialHandle.cs" />
-    <Compile Include="..\ICredential.cs" />
-    <Compile Include="..\PublicEnums.cs" />
-    <Compile Include="..\CredentialAPIException.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="..\..\.editorconfig">
-      <Link>.editorconfig</Link>
-    </None>
-    <None Include="AdysTech.CredentialManager.nuspec" />
-    <None Include="Properties\key.snk" />
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
-       Other similar extension points exist, see Microsoft.Common.targets.
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-  -->
+
 </Project>
