+++ src/AdysTech.CredentialManager/Credential.cs
