+++ src/AdysTech.CredentialManager/CredentialAPIException.cs
