+++ src/AdysTech.CredentialManager/CredentialExtensions.cs
