+++ src/AdysTech.CredentialManager/CredentialManager.cs
