+++ src/AdysTech.CredentialManager/CriticalCredentialHandle.cs
