+++ src/AdysTech.CredentialManager/ICredential.cs
