+++ src/AdysTech.CredentialManager/NativeCode.cs
