+++ src/AdysTech.CredentialManager/Properties/AssemblyInfo.cs
@@ -1,39 +1,3 @@
-using System.Reflection;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
-// General Information about an assembly is controlled through the following
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("AdysTech.CredentialManager")]
-[assembly: AssemblyDescription("Library to store and retreive User credentials from Windows Credential Store")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("AdysTech")]
-[assembly: AssemblyProduct("AdysTech.CredentialManager")]
-[assembly: AssemblyCopyright("Copyright © AdysTech 2020")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible
-// to COM components.  If you need to access a type in this assembly from
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("e96bbbc7-39c4-4be0-ba6c-02b0c5c31577")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.0.0.0")]
-[assembly: AssemblyFileVersion("2.0.0.0")]
-
-//[assembly: InternalsVisibleTo("CredentialManager.Test")] 
-[assembly: InternalsVisibleTo("CredentialManager.Test, PublicKey=0024000004800000940000000602000000240000525341310004000001000100BD5CAC7BD791BBF5443BD0DC379AD9DB0613F97DC6ED2597422DAB9C6535FE3EF53BB6128B977E26E97F59E2ADF8595A4D92219A55F98A548C509392D1CD031C6BC2032941E503A4C1764AFB443D7DF7D47271456636937C6465C8F3AAE63289F51C49E5146A533CB1214110025BB9991550FC0ED1C88E21C5AE42BDED2337C1")]
+[assembly: InternalsVisibleTo("CredentialManager.Test")]
