+++ src/AdysTech.CredentialManager/PublicEnums.cs
