--- tests/CredentialManager.Core.Test/CredentialManager.Core.Test.csproj
@@ -1,28 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-
-    <AssemblyName>CredentialManagerTest</AssemblyName>
-
-    <RootNamespace>CredentialManagerTest</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="..\CredentialManagerTest.cs" Link="CredentialManagerTest.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="1.4.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="1.4.0" />
-    <PackageReference Include="coverlet.collector" Version="1.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\AdysTech.CredentialManager.Core\AdysTech.CredentialManager.Core.csproj" />
-  </ItemGroup>
-
-</Project>
